# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "GuideMessageKind": {
        "description": "문자 한 통이 무엇인가 — 와이어프레임 D1-6 「발송 · 예정」.\n\n회차를 값으로 두는 것이 중요하다. 「일곱째 날」을 시각에서 되계산하면\n진료일이 고쳐질 때 어느 회차였는지를 잃는다.",
        "enum": [
          "GUIDE",
          "CHECK_D7",
          "CHECK_D15",
          "CHECK_D30",
          "RUN_OUT"
        ],
        "title": "GuideMessageKind",
        "type": "string"
      },
...
      "GuidePreview": {
        "additionalProperties": false,
        "description": "스탭·의사 화면의 「환자 화면 미리보기」가 그리는 카드 — KEY-294.\n\n`sections` 만으로는 **나의 목표 · 처방받은 약 · 약별 복용 방법** 세 카드를\n못 그린다. 그 셋은 절이 아니라 처방·검사 기준선에서 나오는 파생이고,\n환자 종점이 이미 짓고 있다. 미리보기가 그것을 못 받아서 「환자가 받는\n그대로」라고 적어 놓고 부분만 보여 주고 있었다 (KEY-286 후속).\n\n**환자 종점과 같은 자리에서 짓는다** — `app/services/patient_guide_view.py`.\n두 곳에서 따로 조립하면 갈라지고, 갈라진 것을 사람은 승인 뒤에야 안다.\n\n새로 나가는 값은 처방·검사 파생과 진료일뿐이다. 링크 토큰·연락처처럼\n미리보기에 없던 개인정보는 여기 담지 않는다.",
        "properties": {
          "guide": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PatientGuideDetailResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "visit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visit"
          }
        },
        "title": "GuidePreview",
        "type": "object"
      },
      "GuideResponse": {
        "additionalProperties": false,
        "properties": {
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientHead"
          },
          "preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GuidePreview"
              },
              {
                "type": "null"
              }
            ]
          },
          "returned_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returned Reason"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionResponse"
            },
            "title": "Sections",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/GuideStatus"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "visit_id": {
            "title": "Visit Id",
            "type": "integer"
          }
        },
        "required": [
          "visit_id",
          "patient",
          "status",
          "version",
          "sections"
        ],
        "title": "GuideResponse",
        "type": "object"
      },
      "GuideSectionKey": {
        "description": "환자 화면의 차례와 같다 — P2 · P3 · P4, 그리고 문자 설정.\n\n`EMERGENCY` 는 `CAUTION` 바로 뒤다. **화면에서 둘은 같은 「주의사항」 탭\n안에 이어 붙는다** — 응급 문장만 따로 탭을 만들면 원장님이 그 탭을 안 열고\n넘길 수 있고, 그 문장은 넘겨도 되는 문장이 아니다.\n\n나눈 까닭은 잠금 단위 때문이다. `locked` 는 섹션 단위라, 🚨 응급 문장을\n지키려고 `caution` 전체를 잠그면 **고칠 수 있어야 할 일반 주의 문구까지\n잠긴다**(와이어프레임 D1-2 — 「🚨 응급 문장만 수정 불가」).",
        "enum": [
          "medication",
          "caution",
          "emergency",
          "life",
          "messages"
        ],
        "title": "GuideSectionKey",
        "type": "string"
      },
      "GuideStatus": {
        "description": "안내문이 지금 어디에 있는가.\n\n**새 이름을 만들지 않는다.** `docs/contracts/patient-visit-api-v1.md` §6 이\n화면 탭과 `detail_status` 를 이미 얼려 두었고, 이 값들이 그대로 그 탭으로\n파생된다. 같은 것을 두 어휘로 부르기 시작하면 어느 쪽이 정본인지 흐려진다.\n\n    STAFF_REVIEW        스탭 확인 중        → 작성 중\n    APPROVAL_PENDING    승인 요청           → 승인 요청\n    SCHEDULED_TO_SEND   발송 예약됨         → 발송 대기\n    APPROVAL_RETURNED   승인 반려           → 보완\n\n승인의 결과가 `APPROVED` 가 아니라 `SCHEDULED_TO_SEND` 인 것이 중요하다 —\n**승인이 곧 발송 예약이다**(와이어프레임 D1-5 「스탭이 발송 버튼을 누르지\n않는다」). 승인과 발송 사이에 사람이 한 번 더 손대는 자리를 만들지 않는다.",
        "enum": [
          "STAFF_REVIEW",
          "APPROVAL_PENDING",
          "SCHEDULED_TO_SEND",
          "APPROVAL_RETURNED"
        ],
        "title": "GuideStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessagePlanRequest": {
        "additionalProperties": false,
        "description": "문자 설정 저장 — 「이 환자만 적용」.",
        "properties": {
          "check_hour": {
            "title": "Check Hour",
            "type": "integer"
          },
          "rounds": {
            "items": {
              "$ref": "#/components/schemas/MessageRound"
            },
            "title": "Rounds",
            "type": "array"
          }
        },
        "required": [
          "check_hour",
          "rounds"
        ],
        "title": "MessagePlanRequest",
        "type": "object"
      },
      "MessagePlanResponse": {
        "additionalProperties": false,
        "properties": {
          "check_hour": {
            "title": "Check Hour",
            "type": "integer"
          },
          "rounds": {
            "items": {
              "$ref": "#/components/schemas/MessageRoundOut"
            },
            "title": "Rounds",
            "type": "array"
          }
        },
        "required": [
          "check_hour",
          "rounds"
        ],
        "title": "MessagePlanResponse",
        "type": "object"
      },
      "MessageRound": {
        "additionalProperties": false,
        "description": "문자 회차 하나 — 와이어프레임 S1-14.\n\n`body` 가 `null` 이면 **기본 문구**다. 빈 문자열도 같은 뜻으로 받는다 —\n스탭이 문구를 다 지운 것은 「빈 문자를 보내라」가 아니라 「기본으로\n되돌려라」이기 때문이다.",
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "days_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Before"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "kind": {
            "$ref": "#/components/schemas/GuideMessageKind"
          }
        },
        "required": [
          "kind",
          "enabled"
        ],
        "title": "MessageRound",
        "type": "object"
      },
      "MessageRoundOut": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "days_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Before"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "fixed": {
            "default": false,
            "title": "Fixed",
            "type": "boolean"
          },
          "kind": {
            "$ref": "#/components/schemas/GuideMessageKind"
          }
        },
        "required": [
          "kind",
          "enabled"
        ],
        "title": "MessageRoundOut",
        "type": "object"
      },
...
      "PatientGender": {
        "enum": [
          "FEMALE",
          "MALE",
          "OTHER",
          "UNKNOWN"
        ],
        "title": "PatientGender",
        "type": "string"
      },
      "PatientGuideDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "drug": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PatientGuideDrugResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "goalSay": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goalsay"
          },
          "goals": {
            "items": {
              "$ref": "#/components/schemas/PatientGuideGoalResponse"
            },
            "title": "Goals",
            "type": "array"
          },
          "how": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "How"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "why": {
            "items": {
              "type": "string"
            },
            "title": "Why",
            "type": "array"
          }
        },
        "title": "PatientGuideDetailResponse",
        "type": "object"
      },
...
      "PatientHead": {
        "additionalProperties": false,
        "description": "승인 화면 머리에 서는 환자. **누구의 안내문인지**를 말한다.\n\n이것이 없으면 원장님은 화면에 이름 없이 뜬 본문을 승인하게 된다. 승인은\n곧 그 환자에게 발송이라, 누구인지 모르고 누르는 자리를 만들면 안 된다.\n\n`birth_date` 와 `age` 를 **함께** 준다 — 계약 §4 가 「동명이인 확인과 계산\n근거를 위해 두 값을 함께 제공한다」로 정해 둔 그대로다. `age` 는 저장값이\n아니라 조회 시점의 현지 날짜로 계산한 읽기 전용 값이다.\n\n`phone` 은 넣지 않는다. 이 화면은 「누구인지」만 알면 되고, 발송 번호는\n서버가 안다. 응답에 실으면 승인할 때마다 전화번호가 화면과 로그를 지난다.",
        "properties": {
          "age": {
            "title": "Age",
            "type": "integer"
          },
          "birth_date": {
            "format": "date",
            "title": "Birth Date",
            "type": "string"
          },
          "gender": {
            "$ref": "#/components/schemas/PatientGender"
          },
          "hospital_patient_no": {
            "title": "Hospital Patient No",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "birth_date",
          "age",
          "gender",
          "hospital_patient_no"
        ],
        "title": "PatientHead",
        "type": "object"
      },
...
      "ReturnRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ReturnRequest",
        "type": "object"
      },
...
      "SectionEditRequest": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "SectionEditRequest",
        "type": "object"
      },
      "SectionResponse": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "edited": {
            "title": "Edited",
            "type": "boolean"
          },
          "key": {
            "$ref": "#/components/schemas/GuideSectionKey"
          },
          "locked": {
            "title": "Locked",
            "type": "boolean"
          },
          "warn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warn"
          }
        },
        "required": [
          "key",
          "body",
          "edited",
          "locked"
        ],
        "title": "SectionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/visits/{visit_id}/guide": {
      "get": {
        "operationId": "read_guide_api_v1_visits__visit_id__guide_get",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Guide",
        "tags": [
          "guides"
        ]
      }
    },
    "/api/v1/visits/{visit_id}/guide/approve": {
      "post": {
        "operationId": "approve_guide_api_v1_visits__visit_id__guide_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Guide",
        "tags": [
          "guides"
        ]
      }
    },
    "/api/v1/visits/{visit_id}/guide/generate": {
      "post": {
        "description": "**고친 문구가 있으면 묻고 멈춘다** — `discard_edits=true` 로 다시 부른다.\n\n다시 만들면 절이 통째로 새로 써진다. 스탭이 이미 바로잡은 문장이 붙어\n있으면 그것이 말없이 사라지므로, 기본값은 **멈추는 쪽**이다(409\n`GUIDE_HAS_EDITS`). 화면이 「고친 것을 버리고 다시 만들까요」를 묻고, 사람이\n그렇다고 하면 이 값을 켜서 다시 부른다 (이희진 님 `#221` ①).",
        "operationId": "generate_guide_api_v1_visits__visit_id__guide_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "discard_edits",
            "required": false,
            "schema": {
              "default": false,
              "title": "Discard Edits",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Guide",
        "tags": [
          "guides"
        ]
      }
    },
...
    "/api/v1/visits/{visit_id}/guide/messages": {
      "get": {
        "description": "이 진료의 문자 회차 설정 — 와이어프레임 S1-14.\n\n한 번도 안 만진 진료도 기본값으로 답한다. 「설정이 없다」와 「기본값이다」를\n화면이 가를 이유가 없다.",
        "operationId": "read_message_plan_api_v1_visits__visit_id__guide_messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Message Plan",
        "tags": [
          "guides"
        ]
      },
      "put": {
        "description": "문자 설정을 저장한다 — 「이 환자만 적용」.\n\n**한 판을 통째로 받는다.** 회차 하나씩 PATCH 로 받으면, 중간에 끊겼을 때\n「보름 뒤는 껐는데 한 달 뒤는 안 켜진」 반쪽 상태가 남는다.",
        "operationId": "save_message_plan_api_v1_visits__visit_id__guide_messages_put",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessagePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Message Plan",
        "tags": [
          "guides"
        ]
      }
    },
    "/api/v1/visits/{visit_id}/guide/return": {
      "post": {
        "operationId": "return_guide_api_v1_visits__visit_id__guide_return_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Return Guide",
        "tags": [
          "guides"
        ]
      }
    },
    "/api/v1/visits/{visit_id}/guide/sections/{key}": {
      "patch": {
        "operationId": "edit_section_api_v1_visits__visit_id__guide_sections__key__patch",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Edit Section",
        "tags": [
          "guides"
        ]
      }
    },
    "/api/v1/visits/{visit_id}/guide/submit": {
      "post": {
        "description": "스탭이 확인을 마치고 의사에게 넘긴다 — 와이어프레임 S1-11.\n\n이 자리가 없어서 안내문이 만들어지자마자 원장님 목록에 떴다.\n승인은 여전히 의사만 한다 (`/guide/approve`).",
        "operationId": "submit_guide_api_v1_visits__visit_id__guide_submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Guide",
        "tags": [
          "guides"
        ]
      }
    },
    "/api/v1/visits/{visit_id}/guide/unapprove": {
      "post": {
        "description": "승인을 거둔다 — 승인했는데 잘못된 것을 발견했을 때.\n\n이미 나간 문자가 있으면 거두지 않는다. 예약된 문자는 꺼진다.",
        "operationId": "unapprove_guide_api_v1_visits__visit_id__guide_unapprove_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "title": "Visit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Unapprove Guide",
        "tags": [
          "guides"
        ]
      }
    },
...
  }
}